+++ test/OpenTelemetry.Exporter.Jaeger.Tests/Implementation/JaegerActivityConversionTest.cs
@@ -20,7 +20,7 @@ using System.Linq;
 
 using OpenTelemetry.Exporter.Jaeger.Implementation;
 using OpenTelemetry.Resources;
-
+using OpenTelemetry.Trace;
 using Xunit;
 
 namespace OpenTelemetry.Exporter.Jaeger.Tests.Implementation
@@ -396,26 +396,21 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests.Implementation
             Assert.Empty(jaegerSpan.Tags.Where(t => t.Key == "peer.service"));
         }
 
-        [Fact]
-        public void JaegerActivityConverterTest_GenerateJaegerSpan_RemoteEndpointResolutionPriority()
+        [Theory]
+        [MemberData(nameof(RemoteEndpointPriorityTestCase.GetTestCases), MemberType = typeof(RemoteEndpointPriorityTestCase))]
+        public void JaegerActivityConverterTest_GenerateJaegerSpan_RemoteEndpointResolutionPriority(RemoteEndpointPriorityTestCase testCase)
         {
             // Arrange
-            var span = CreateTestActivity(
-                additionalAttributes: new Dictionary<string, object>
-                {
-                    ["http.host"] = "DiscardedRemoteServiceName",
-                    ["peer.service"] = "RemoteServiceName",
-                    ["peer.hostname"] = "DiscardedRemoteServiceName",
-                });
+            var activity = CreateTestActivity(additionalAttributes: testCase.RemoteEndpointAttributes);
 
             // Act
-            var jaegerSpan = span.ToJaegerSpan();
+            var jaegerSpan = activity.ToJaegerSpan();
 
             // Assert
             var tags = jaegerSpan.Tags.Where(t => t.Key == "peer.service");
             Assert.Single(tags);
             var tag = tags.First();
-            Assert.Equal("RemoteServiceName", tag.VStr);
+            Assert.Equal(testCase.ExpectedResult, tag.VStr);
         }
 
         internal static Activity CreateTestActivity(
@@ -509,6 +504,132 @@ namespace OpenTelemetry.Exporter.Jaeger.Tests.Implementation
             activity.Stop();
 
             return activity;
+        }
+
+        public class RemoteEndpointPriorityTestCase
+        {
+            public string Name { get; set; }
+
+            public string ExpectedResult { get; set; }
+
+            public Dictionary<string, object> RemoteEndpointAttributes { get; set; }
+
+            public static IEnumerable<object[]> GetTestCases()
+            {
+                yield return new object[]
+                {
+                    new RemoteEndpointPriorityTestCase
+                    {
+                        Name = "Highest priority name = net.peer.name",
+                        ExpectedResult = "RemoteServiceName",
+                        RemoteEndpointAttributes = new Dictionary<string, object>
+                        {
+                            ["http.host"] = "DiscardedRemoteServiceName",
+                            ["net.peer.name"] = "RemoteServiceName",
+                            ["peer.hostname"] = "DiscardedRemoteServiceName",
+                        },
+                    },
+                };
+
+                yield return new object[]
+                {
+                    new RemoteEndpointPriorityTestCase
+                    {
+                        Name = "Highest priority name = SemanticConventions.AttributePeerService",
+                        ExpectedResult = "RemoteServiceName",
+                        RemoteEndpointAttributes = new Dictionary<string, object>
+                        {
+                            [SemanticConventions.AttributePeerService] = "RemoteServiceName",
+                            ["http.host"] = "DiscardedRemoteServiceName",
+                            ["net.peer.name"] = "DiscardedRemoteServiceName",
+                            ["net.peer.port"] = "1234",
+                            ["peer.hostname"] = "DiscardedRemoteServiceName",
+                        },
+                    },
+                };
+
+                yield return new object[]
+                {
+                    new RemoteEndpointPriorityTestCase
+                    {
+                        Name = "Only has net.peer.name and net.peer.port",
+                        ExpectedResult = "RemoteServiceName:1234",
+                        RemoteEndpointAttributes = new Dictionary<string, object>
+                        {
+                            ["net.peer.name"] = "RemoteServiceName",
+                            ["net.peer.port"] = "1234",
+                        },
+                    },
+                };
+
+                yield return new object[]
+                {
+                    new RemoteEndpointPriorityTestCase
+                    {
+                        Name = "net.peer.port is an int",
+                        ExpectedResult = "RemoteServiceName:1234",
+                        RemoteEndpointAttributes = new Dictionary<string, object>
+                        {
+                            ["net.peer.name"] = "RemoteServiceName",
+                            ["net.peer.port"] = 1234,
+                        },
+                    },
+                };
+
+                yield return new object[]
+                {
+                    new RemoteEndpointPriorityTestCase
+                    {
+                        Name = "Has net.peer.name and net.peer.port",
+                        ExpectedResult = "RemoteServiceName:1234",
+                        RemoteEndpointAttributes = new Dictionary<string, object>
+                        {
+                            ["http.host"] = "DiscardedRemoteServiceName",
+                            ["net.peer.name"] = "RemoteServiceName",
+                            ["net.peer.port"] = "1234",
+                            ["peer.hostname"] = "DiscardedRemoteServiceName",
+                        },
+                    },
+                };
+
+                yield return new object[]
+                {
+                    new RemoteEndpointPriorityTestCase
+                    {
+                        Name = "Has net.peer.ip and net.peer.port",
+                        ExpectedResult = "1.2.3.4:1234",
+                        RemoteEndpointAttributes = new Dictionary<string, object>
+                        {
+                            ["http.host"] = "DiscardedRemoteServiceName",
+                            ["net.peer.ip"] = "1.2.3.4",
+                            ["net.peer.port"] = "1234",
+                            ["peer.hostname"] = "DiscardedRemoteServiceName",
+                        },
+                    },
+                };
+
+                yield return new object[]
+                {
+                    new RemoteEndpointPriorityTestCase
+                    {
+                        Name = "Has net.peer.name, net.peer.ip, and net.peer.port",
+                        ExpectedResult = "RemoteServiceName:1234",
+                        RemoteEndpointAttributes = new Dictionary<string, object>
+                        {
+                            ["http.host"] = "DiscardedRemoteServiceName",
+                            ["net.peer.name"] = "RemoteServiceName",
+                            ["net.peer.ip"] = "1.2.3.4",
+                            ["net.peer.port"] = "1234",
+                            ["peer.hostname"] = "DiscardedRemoteServiceName",
+                        },
+                    },
+                };
+            }
+
+            public override string ToString()
+            {
+                return this.Name;
+            }
         }
     }
 }